Se termina la parte del login auth validaciones

+++ src/controllers/ProjectController.ts
@@ -7,7 +7,9 @@ export class ProjectController{
     static createProject = async(req: Request, res: Response)=>{
 
         const project = new Project(req.body)
-        
+
+        // Asigna un manager
+        project.manager = req.user._id
         try {
             await project.save()
             // await Project.create(req.body)
@@ -20,7 +22,11 @@ export class ProjectController{
     static getAllProjects = async(req: Request, res: Response)=>{
 
         try {
-            const projects = await Project.find({}).populate('tasks')
+            const projects = await Project.find({
+                $or:[
+                    {manager: {$in: req.user._id}}
+                ]
+            }) //.populate('tasks')
             res.json(projects)
 
         } catch (error) {
@@ -32,13 +38,20 @@ export class ProjectController{
         
         const {id} = req.params
         try {
-            const projects = await (await Project.findById(id).populate('tasks'))
-            if(!projects)
+            const project = await (await Project.findById(id).populate('tasks'))
+            if(!project)
             {
                 const error = new Error('Project not found')
                 return res.status(404).json({error: error.message})
             }
-            res.json(projects)
+                
+            if(project.manager.toString() !== req.user._id.toString())
+            {
+                const error = new Error('Action not valid')
+                return res.status(404).json({error: error.message})
+            }
+            
+            res.json(project)
 
         } catch (error) {
             return res.status(500).json({error: error.message})
@@ -56,6 +69,13 @@ export class ProjectController{
                 const error = new Error('Project not found')
                 return res.status(404).json({error: error.message})
             } 
+ 
+            if(project.manager.toString() !== req.user._id.toString())
+            {
+                const error = new Error('only the manager can update a project')
+                return res.status(404).json({error: error.message})
+            }
+
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -80,6 +100,12 @@ export class ProjectController{
                 return res.status(404).json({error: error.message})
             } 
             
+            if(project.manager.toString() !== req.user._id.toString())
+            {
+                const error = new Error('only the manager can update a project')
+                return res.status(404).json({error: error.message})
+            }
+
             await project?.deleteOne()
             
             res.send('Project deleted successfully')
